More refactoring Excel of Calculation Function Unit Tests
