+++ tests/PhpSpreadsheetTests/Calculation/LookupRefTest.php
@@ -20,70 +20,6 @@ class LookupRefTest extends TestCase
     }
 
     /**
-     * @dataProvider providerMATCH
-     *
-     * @param mixed $expectedResult
-     */
-    public function testMATCH($expectedResult, ...$args)
-    {
-        $result = LookupRef::MATCH(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerMATCH()
-    {
-        return require 'data/Calculation/LookupRef/MATCH.php';
-    }
-
-    /**
-     * @dataProvider providerINDEX
-     *
-     * @param mixed $expectedResult
-     */
-    public function testINDEX($expectedResult, ...$args)
-    {
-        $result = LookupRef::INDEX(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerINDEX()
-    {
-        return require 'data/Calculation/LookupRef/INDEX.php';
-    }
-
-    /**
-     * @dataProvider providerCOLUMNS
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOLUMNS($expectedResult, ...$args)
-    {
-        $result = LookupRef::COLUMNS(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerCOLUMNS()
-    {
-        return require 'data/Calculation/LookupRef/COLUMNS.php';
-    }
-
-    /**
-     * @dataProvider providerROWS
-     *
-     * @param mixed $expectedResult
-     */
-    public function testROWS($expectedResult, ...$args)
-    {
-        $result = LookupRef::ROWS(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerROWS()
-    {
-        return require 'data/Calculation/LookupRef/ROWS.php';
-    }
-
-    /**
      * @dataProvider providerFormulaText
      *
      * @param mixed $expectedResult
