+++ tests/data/Calculation/LookupRef/CHOOSE.php
@@ -1,0 +1,28 @@
+<?php
+
+return [
+    [
+        "brown",
+        4, "red", "blue", "green", "brown",
+    ],
+    [
+        "blue",
+        2, "red", "blue", "green", "brown",
+    ],
+    [
+        "green",
+        3, "red", "blue", "green", "brown",
+    ],
+    [
+        "red",
+        1, "red", "blue", "green", "brown",
+    ],
+    [
+        "#VALUE!",
+        5, "red", "blue", "green", "brown",
+    ],
+    [
+        "#VALUE!",
+        0, "red", "blue", "green", "brown",
+    ],
+];
